Refactor / simplify / fix C++ exceptions
