+++ cpp/src/Ice/Instance.cpp
@@ -113,7 +113,7 @@ namespace
         {
             try
             {
-                Exception::ice_enableStackTraceCollection();
+                LocalException::ice_enableStackTraceCollection();
             }
             catch (const std::exception& ex)
             {
