+++ cpp/src/Ice/OutgoingResponse.cpp
@@ -20,9 +20,9 @@ namespace
 {
     inline string toString(const Exception& ex)
     {
-        // Includes the stack trace when available.
+        // Includes the stack trace when available (local exceptions only).
         ostringstream os;
-        os << ex;
+        ex.ice_print(os);
         return os.str();
     }
 
@@ -151,13 +151,6 @@ namespace
             exceptionDetails = toString(ex);
             unknownExceptionMessage = createUnknownExceptionMessage(exceptionId, ex.what());
             replyStatus = ReplyStatus::UnknownLocalException;
-        }
-        catch (const Exception& ex)
-        {
-            exceptionId = ex.ice_id();
-            exceptionDetails = toString(ex);
-            unknownExceptionMessage = createUnknownExceptionMessage(exceptionId, ex.what());
-            replyStatus = ReplyStatus::UnknownException;
         }
         catch (const std::exception& ex)
         {
