+++ cpp/src/IceBox/ServiceManagerI.cpp
@@ -441,7 +441,7 @@ IceBox::ServiceManagerI::start()
         {
             try
             {
-                Exception::ice_enableStackTraceCollection();
+                LocalException::ice_enableStackTraceCollection();
             }
             catch (const std::exception& ex)
             {
